Consolidate pool and raw SQL into db.server, delete scripts/db.ts
